+++ src/registers/databaseTables.php
@@ -70,6 +70,7 @@
         ["Name" => "wormholeprice", "Type" => "BIGINT"], 
         ["Name" => "maxpochvenvolume", "Type" => "BIGINT"], 
         ["Name" => "pochvenprice", "Type" => "BIGINT"], 
+        ["Name" => "minimumprice", "Type" => "BIGINT"], 
         ["Name" => "collateralpremium", "Type" => "NUMERIC(8,4)"],
         ["Name" => "highcollateralcutoff", "Type" => "BIGINT"], 
         ["Name" => "highcollateralpenalty", "Type" => "BIGINT"], 
@@ -88,10 +89,12 @@
         ["Name" => "end", "Type" => "BIGINT"],
         ["Name" => "basepriceoverride", "Type" => "BIGINT", "Special" => "DEFAULT NULL"],
         ["Name" => "gatepriceoverride", "Type" => "BIGINT", "Special" => "DEFAULT NULL"],
+        ["Name" => "minimumpriceoverride", "Type" => "BIGINT", "Special" => "DEFAULT NULL"],
         ["Name" => "pricemodel", "Type" => "TEXT"],
         ["Name" => "collateralpremiumoverride", "Type" => "NUMERIC(8,4)", "Special" => "DEFAULT NULL"],
         ["Name" => "maxvolumeoverride", "Type" => "BIGINT", "Special" => "DEFAULT NULL"],
         ["Name" => "maxcollateraloverride", "Type" => "BIGINT", "Special" => "DEFAULT NULL"],
+        ["Name" => "disablehighcollateral", "Type" => "TINYINT"], 
         ["Name" => "", "Type" => "", "Special" => "CONSTRAINT PK_ROUTES_MUST_BE_UNIQUE PRIMARY KEY (start, end)"],
         ["Name" => "", "Type" => "", "Special" => "CONSTRAINT CHK_FIXED_NEEDS_PRICE CHECK (pricemodel != 'Fixed' OR basepriceoverride IS NOT NULL)"]
     );